contract_method: add in_signature tuple[] support
There is one caveat: we were abusing that when having a tuple
the names of the elements do not repeat and we were using that
in the fixed arguments. Using AvatarAddress for tuple or tuple[]
is no longer supported in the fixed_arguments. You must put the
avatar value in the tuple or tuple[] manualy.